Moved the calendar gutter into a separate component
